+++ App.xaml.cs
@@ -7,6 +7,7 @@ using System.IO.Abstractions;
 using System.Collections.ObjectModel;
 using System.IO;
 using System.Text.Json;
+using System.Threading;
 using System.Threading.Tasks;
 
 using Microsoft.Extensions.DependencyInjection;
@@ -20,6 +21,7 @@ using WifftOCR.Services.Consumers;
 using WifftOCR.Loggers;
 using WifftOCR.DataModels;
 using WifftOCR.Interfaces;
+using WifftOCR.Providers;
 using WifftOCR.Services;
 using WifftOCR.ViewModels;
 using WifftOCR.Views;
@@ -28,11 +30,15 @@ namespace WifftOCR
 {
     public partial class App : Application
     {
+        private Mutex _mutex;
+
         private MainWindow m_window;
 
         public const string SETTINGS_LOCATION_URI = "ms-appdata:///roaming/settings.json";
+        public const string LOG_FILE_LOCATION_URI = "ms-appdata:///roaming/system.log";
 
         public IHost Host { get; }
+        public ILoggerFactory LoggerFactory { get; private set;  }
 
         public static T GetService<T>() where T : class
         {
@@ -59,39 +65,62 @@ namespace WifftOCR
                     logging.SetMinimumLevel(LogLevel.Debug)
                         .AddFilter("Microsoft", LogLevel.Warning)
                         .AddFilter("Microsoft.Hosting.Lifetime", LogLevel.Warning);
-                        /*.AddConsole(options => options.FormatterName = "wifftFormatter")
-                        .AddConsoleFormatter<CustomFormatter, CustomFormatter.Options>(options => options.CustomPrefix = "[WifftOCR] ");*/
                 })
                 .ConfigureServices((context, services) => {
                     services.AddSingleton<IFileSystem, FileSystem>();
                     services.AddSingleton<ISettingsService, SettingsService>();
+                    services.AddSingleton<IFileLoggerService, FileLoggerService>();
 ;
-                    services.AddHostedService<OcrServiceConsumer>();
-                    services.AddScoped<IScopedProcessingService, OcrService>();
+                    services.AddHostedService<OcrRecorderServiceConsumer>();
+                    services.AddScoped<IScopedProcessingService, OcrRecorderService>();
 
                     services.AddTransient<ShellPage>();
                     services.AddTransient<ShellViewModel>();
 
                     services.AddTransient<WelcomePage>();
+                    services.AddTransient<WelcomePageViewModel>();
 
                     services.AddTransient<CaptureAreasPage>();
                     services.AddTransient<CaptureAreasViewModel>();
 
                     services.AddTransient<SettingsPage>();
                     services.AddTransient<SettingsViewModel>();
-
-                    services.AddTransient<XamlLogViewerLogger>();
                 })
                 .Build();
         }
 
         protected override async void OnLaunched(LaunchActivatedEventArgs args)
         {
+            _mutex = new Mutex(true, "WifftOCR_SingleInstanceMutex", out bool isNewInstance);
+            if (!isNewInstance) {
+                m_window = Window.Current as MainWindow;
+                m_window?.Activate();
+
+                Current.Exit();
+
+                return;
+            }
+
+            base.OnLaunched(args);
+
             m_window = new MainWindow();
             m_window.Activate();
             MainWindow.EnsurePageIsSelected();
 
             await CheckIfConfigFileExists();
+            await CheckIfLogFileExists();
+
+            BuildLoggerFactory();
+
+            _mutex.ReleaseMutex();
+        }
+
+        private void BuildLoggerFactory()
+        {
+            ILoggerFactory loggerFactory = Host.Services.GetService<ILoggerFactory>();
+            loggerFactory.AddProvider(new FileLoggerProvider());
+
+            LoggerFactory = loggerFactory;
         }
 
         private static async Task CheckIfConfigFileExists()
@@ -100,7 +129,6 @@ namespace WifftOCR
 
             try {
                 configFile = await StorageFile.GetFileFromApplicationUriAsync(new Uri(SETTINGS_LOCATION_URI));
-                Console.WriteLine(configFile.Path);
             } catch (FileNotFoundException) {
                 StorageFolder folder = ApplicationData.Current.RoamingFolder;
 
@@ -122,6 +150,24 @@ namespace WifftOCR
 
             using Stream stream = await file.OpenStreamForWriteAsync();
             await JsonSerializer.SerializeAsync(stream, settings);
+        }
+
+        private static async Task CheckIfLogFileExists()
+        {
+            StorageFile logFile;
+
+            try {
+                logFile = await StorageFile.GetFileFromApplicationUriAsync(new Uri(LOG_FILE_LOCATION_URI));
+            } catch (FileNotFoundException) {
+                StorageFolder folder = ApplicationData.Current.RoamingFolder;
+
+                await CreateLogFile(folder);
+            }
+        }
+
+        private static async Task CreateLogFile(StorageFolder folder)
+        {
+            await folder.CreateFileAsync("system.log", CreationCollisionOption.OpenIfExists);
         }
     }
 }