+++ Services/FileLoggerService.cs
@@ -5,18 +5,16 @@
 using System;
 using System.IO;
 using System.IO.Abstractions;
-using System.Text.Json;
 using System.Threading;
 using System.Threading.Tasks;
 
 using Windows.Storage;
 
 using WifftOCR.Interfaces;
-using WifftOCR.DataModels;
 
 namespace WifftOCR.Services
 {
-    internal class SettingsService : ISettingsService, IDisposable
+    internal class FileLoggerService : IFileLoggerService, IDisposable
     {
         private readonly SemaphoreSlim _asyncLock = new(1, 1);
         private readonly IFileSystem _fileSystem;
@@ -30,11 +28,11 @@ namespace WifftOCR.Services
         public event EventHandler FileChanged;
 
         #nullable enable
-        public SettingsService(IFileSystem? fileSystem = null)
+        public FileLoggerService(IFileSystem? fileSystem = null)
         {
             _fileSystem = fileSystem;
 
-            _settingsFilePath = StorageFile.GetFileFromApplicationUriAsync(new Uri(App.SETTINGS_LOCATION_URI))
+            _settingsFilePath = StorageFile.GetFileFromApplicationUriAsync(new Uri(App.LOG_FILE_LOCATION_URI))
                 .GetAwaiter()
                 .GetResult()
                 .Path;
@@ -50,50 +48,69 @@ namespace WifftOCR.Services
             }
         }
 
-        public async Task<bool> WriteToFileAsync(Settings settings)
+        public async Task<bool> WriteToFileAsync(string logLine)
         {
             try {
                 await _asyncLock.WaitAsync();
-                _fileSystemWatcher.EnableRaisingEvents = false;
 
-                StorageFile file = await StorageFile.GetFileFromApplicationUriAsync(new Uri(App.SETTINGS_LOCATION_URI));
+                StorageFile file = await StorageFile.GetFileFromApplicationUriAsync(new Uri(App.LOG_FILE_LOCATION_URI));
 
-                using Stream stream = await file.OpenStreamForWriteAsync();
-                stream.SetLength(0);
-
-                JsonSerializerOptions options = new() {
-                    WriteIndented = true
-                };
-
-                await JsonSerializer.SerializeAsync(stream, settings, options);
+                using StreamWriter writer = new(await file.OpenStreamForWriteAsync());
+                await writer.WriteLineAsync(logLine);
+                writer.Close();
             } catch {
                 return false;
             } finally {
-                _fileSystemWatcher.EnableRaisingEvents = true;
                 _asyncLock.Release();
             }
-            
+
+            return true;
+        }
+
+        public async Task<bool> ClearFileAsync()
+        {
+            try {
+                await _asyncLock.WaitAsync();
+
+                StorageFile file = await StorageFile.GetFileFromApplicationUriAsync(new Uri(App.LOG_FILE_LOCATION_URI));
+
+                Stream stream = await file.OpenStreamForWriteAsync();
+                stream.SetLength(0);
+
+                using StreamWriter writer = new(stream);
+                await writer.WriteAsync(string.Empty);
+                
+                writer.Close();
+            } catch {
+                return false;
+            } finally {
+                _asyncLock.Release();
+            }
+
             return true;
         }
 
         #nullable enable
-        public async Task<Settings?> ReadFromFileAsync()
+        public async Task<string> ReadFromFileAsync()
         {
-            try {
-                StorageFile file = await StorageFile.GetFileFromApplicationUriAsync(new Uri(App.SETTINGS_LOCATION_URI));
-                using StreamReader reader = new(await file.OpenStreamForReadAsync());
+            StorageFile file = await StorageFile.GetFileFromApplicationUriAsync(new Uri(App.LOG_FILE_LOCATION_URI));
+            using StreamReader reader = new(await file.OpenStreamForReadAsync());
 
-                return JsonSerializer.Deserialize<Settings>(await reader.ReadToEndAsync());
-            } catch (IOException e) {
-                Console.WriteLine(e.Message);
+            string lines = reader.ReadToEnd();
 
-                return null;
-            }
+            reader.Close();
+
+            return lines;
         }
 
-        public static async Task<Settings?> ReadSettingsForOcrServiceAsync()
+        public static async Task<bool> WriteLogFileForServiceProviderAsync(string logLine)
         {
-            return await (new SettingsService(null)).ReadFromFileAsync();
+            return await (new FileLoggerService(null)).WriteToFileAsync(logLine);
+        }
+
+        public static async Task<string> ReadLogFileForServiceProviderAsync()
+        {
+            return await (new FileLoggerService(null)).ReadFromFileAsync();
         }
 
         private void FileSystemWatcher_Changed(object sender, FileSystemEventArgs e)