+++ Views/WelcomePage.xaml.cs
@@ -2,41 +2,56 @@
 // Wifft licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System;
-
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Logging;
 using Microsoft.UI.Dispatching;
 using Microsoft.UI.Xaml;
 using Microsoft.UI.Xaml.Controls;
-
-
-using WifftOCR.Providers;
+using Microsoft.UI.Xaml.Media;
+using WifftOCR.ViewModels;
 
 namespace WifftOCR.Views
 {
     public sealed partial class WelcomePage : Page
     {
+        private DispatcherQueueTimer _timer;
+
+        public WelcomePageViewModel ViewModel { get; set; }
+
         public WelcomePage()
         {
             InitializeComponent();
 
-            IServiceProvider serviceProvider = App.GetInstance().Host.Services;
+            ViewModel = App.GetService<WelcomePageViewModel>();
+            DataContext = ViewModel;
 
-            DispatcherQueue.GetForCurrentThread().TryEnqueue(() => {
-                ILoggerFactory loggerFactory = serviceProvider.GetService<ILoggerFactory>();
-                loggerFactory.AddProvider(new XamlLogViewerProvider(LogViewer));
-            });
+            CompositionTarget.Rendering += CompositionTarget_Rendering;
         }
 
         private async void StartButton_Click(object sender, RoutedEventArgs e)
         {
-            await App.GetInstance().Host.StartAsync();
+            Button button = sender as Button;
+            button.IsEnabled = false;
+
+            StopButton.IsEnabled = true;   
+
+            await ViewModel.StartOcrRecorderService();
         }
 
         private async void StopButton_Click(object sender, RoutedEventArgs e)
         {
-            await App.GetInstance().Host.StopAsync();
+            Button button = sender as Button;
+            button.IsEnabled = false;
+
+            StartButton.IsEnabled = true;
+
+            await ViewModel.StopOcrRecorderService();
+        }
+
+        private void CompositionTarget_Rendering(object sender, object e)
+        {
+            if (ViewModel.OcrRecordingServiceRunning) {
+                ViewModel.ReadLogFile();
+                LogViewer.Text = ViewModel.LogLines;
+            }
         }
     }
 }